+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Fixed
 - [Everyone] If the new version information cannot be retrieved from GitLab, the what's new dialog is not shown. (@bbakos)
+- [Student] Fix hidden skills only appearing after page reload.
 
 ## [2.2.7]